185336568-log-link-requests (#3142)
This makes logging the responsibility of OauthClientConnection. Logging will happen by default going forward without needing to reimplement it for every API. The log record is returned as part of the result from the client so that it can be associated with our db records.

Additionally, the request is logged before the external API is invoked. That way we will retain a record of the attempted request if something goes wrong.

To associate the log records with the correct initiator, we pass the credentials into the OauthClientConnection. This presented the opportunity to  simplify the configuration interface for OauthClientConnection. It now takes a credential object and reads its configuration from that.

Functional changes:
* logs all LINK requests
* replaces previous logging implementation for MCI requests
* default json content type on get requests
* side-effect, HmisExternalApis::AcHmis::DataWarehouseApi will now also log requests