Use SubmitForm for creating outgoing referral, support custom data elements (#4389)
Deprecate CreateOutgoingReferralPostingPayload
Update SubmitForm to support creating ReferralPostings, with test coverage in submit_form_spec
Add REFERRAL role for definition. The Referral form represents the form that is collected when a referral is sent. It is not (currenlty) used for editing/updating that referral. Updating the referral status still happens via the UpdateReferralPosting mutations.
Make unit_type_id optional on referrals. It will still be required on input for ac environment because it is marked required on their referral form.
Resolve Custom Data Elements on ReferralPosting, and allow updating them via SubmitForm
Resolve source referral on Enrollment, to support linking to it